Dependency Resolver (#12)
* Adds functionality for Broken mods folder

Allows the updater to moves mods to and from the Broken mods folder (also made by VRCMelonAssistant)
It will move any mods that are flagged as broken by the Remote API to the Broken folder as long as the local version is not newer then the remote version. (Allows modders to locally test new versions)
Will move any mods that are in the Broken folder to Mods if the remote API says it is Approved and remote version is > local (Allows someone to manually move a mod to the broken folder if not flagged by the API)
Has a MelonPref to disable the moving To/From Broken folder completely

* Update VRCModUpdaterCore.cs

Quick Fix to make sure dir exists and apparently didn't submit version with MelonPref

* Dependency Resolver

Added option to add missing required and optional Dependencies. Will only count those found in the RemoteAPI

Retains previous behavior where if the mod dependency is in the broken folder and version number is the same as RemoteAPI, it will not auto move.

Co-authored-by: Hugo "Slaynash" Flores <hugo.flores@etu.univ-lyon1.fr>