Bug fixes and Update Message Popup (#17)
* Adds functionality for Broken mods folder

Allows the updater to moves mods to and from the Broken mods folder (also made by VRCMelonAssistant)
It will move any mods that are flagged as broken by the Remote API to the Broken folder as long as the local version is not newer then the remote version. (Allows modders to locally test new versions)
Will move any mods that are in the Broken folder to Mods if the remote API says it is Approved and remote version is > local (Allows someone to manually move a mod to the broken folder if not flagged by the API)
Has a MelonPref to disable the moving To/From Broken folder completely

* Update VRCModUpdaterCore.cs

Quick Fix to make sure dir exists and apparently didn't submit version with MelonPref

* Dependency Resolver

Added option to add missing required and optional Dependencies. Will only count those found in the RemoteAPI

Retains previous behavior where if the mod dependency is in the broken folder and version number is the same as RemoteAPI, it will not auto move.

* APIv1, Bug fixes, Popup, Caching

Updated for APIv1
  Matching upstream
  Depresolver checking redirect filename
Dependency Resolver BugFix
  Previously would previously newly updated mods from loading till next reload, uses silly Ceil stuff now to fix that to
Popup
  Now has a popup to list what was changed, disableable in Mod Settings
Cachine
   Now can cache your local mod folders for quicker loading, maybe?

* Update VersionUtils.cs

Less spammy

* Bugfix and Popup

Dependency Resolver has been changed to check for redirect URL and get the file name from that, this is needed with the change to API v1

Dependency Resolver would previously prevent newly updated mods from loading till next reload, uses silly Ceil stuff now to fix that

Now has a popup to list what was changed, disableable in Mod Settings

* Update ModDetail.cs

Removed changes for caching

Co-authored-by: Hugo "Slaynash" Flores <hugo.flores@etu.univ-lyon1.fr>